src/requirements.txt
torch
torchvision
torchaudio
opencv-python
mediapipe
numpy
pillow
scikit-learn
matplotlib